pyproject.toml
[project]
name = "model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=4.1.0",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "uvicorn>=0.34.3",
]
